Legge til validering for dokumentasjonskrav UUID for å hindre SSRF

+++ src/pages/api/documentation/file/delete.ts
@@ -7,6 +7,7 @@ import {
   getSoknadOnBehalfOfToken,
 } from "../../../../auth.utils";
 import { logRequestError } from "../../../../error.logger";
+import { isValidUUID } from "../../../../utils/uuid.utils";
 
 export interface IDeleteFileBody {
   uuid: string;
@@ -25,6 +26,11 @@ async function deleteFileHandler(req: NextApiRequest, res: NextApiResponse) {
   }
 
   const { uuid, dokumentkravId, filsti } = req.body;
+
+  if (!isValidUUID(uuid)) {
+    throw Error("Ugyldig uuid");
+  }
+
   const soknadOnBehalfOfToken = await getSoknadOnBehalfOfToken(session);
   const mellomlagringOnBehalfOfToken = await getMellomlagringOnBehalfOfToken(session);
 
@@ -73,6 +79,10 @@ async function deleteFileFromDPSoknad(
   onBehalfOfToken: string,
   filsti: string
 ) {
+  if (!isValidUUID(uuid)) {
+    throw Error("Ugyldig uuid");
+  }
+
   const url = `${process.env.API_BASE_URL}/soknad/${uuid}/dokumentasjonskrav/${dokumentkravId}/fil/${filsti}`;
   return fetch(url, {
     method: "DELETE",
